Move disable submit code to NOD instead of HLR
